ai_models/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
PyMuPDF>=1.24.0
pytesseract>=0.3.10
